backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
